requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.6.0
Flask-Caching==2.3.1
python-dotenv==1.2.1
bleach==6.3.0
qrcode[pil]==8.2
gunicorn==24.1.1
eventlet==0.37.0
cryptography==46.0.3
requests==2.32.5
email_validator==2.3.0
mysql-connector-python==9.5.0
redis==7.1.0
